+++ UART/src/transfer.c
@@ -8,7 +8,7 @@ size_t UART_Write(UART* uart, void* _data, size_t size)
 	uint8_t* data    = _data;
 	size_t   written = 0;
 
-	uart->Registers->StatusB &= (uint8_t) ~(1 << UDRE0); // Disable TX interrupt
+	uart->Registers->StatusB &= ~(1 << UDRE0); // Disable TX interrupt
 
 	uint8_t* bufferElement = FifoBuffer_Add(&uart->TxBuffer);
 	while (bufferElement)
@@ -29,7 +29,7 @@ size_t UART_Read(UART* uart, void* _data, size_t size)
 	uint8_t* data = _data;
 	size_t   read = 0;
 
-	uart->Registers->StatusB &= (uint8_t) ~(1 << RXC0); // Disable RX interrupt
+	uart->Registers->StatusB &= ~(1 << RXC0); // Disable RX interrupt
 
 	uint8_t* bufferElement = FifoBuffer_Remove(&uart->RxBuffer);
 	while (bufferElement)
@@ -45,4 +45,7 @@ size_t UART_Read(UART* uart, void* _data, size_t size)
 	return read;
 }
 
-bool UART_IsBusy(UART* uart) { return (uart->Registers->StatusB & (1 << UDRE0)); /* Active if TX interrupt is enabled */ }
+bool UART_IsBusy(UART* uart)
+{
+	return (uart->Registers->StatusB & (1 << UDRE0)); // Active if TX interrupt is enabled
+}
